Added an update header with the number of offers
